chore(veriForm): removed hostel and roll feilds
