+++ components/footer.tsx
@@ -11,8 +11,18 @@ export default function footer({}: Props) {
       <Separator />
       <div className="px-20">
         <div className="mt-5 flex justify-center gap-3">
-          <FaDiscord size={25} />
-          <MdEmail size={25} />
+          <a
+            href="https://discord.gg/bEMrcuW3W3"
+            target="_blank"
+            rel="noopener noreferrer"
+            title="Join our Discord">
+            <FaDiscord size={25} />
+          </a>
+          <a
+            href="mailto:team@weride.live"
+            title="Mail us your Query">
+            <MdEmail size={25} />
+          </a>
         </div>
       </div>
     </div>
